# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Customer {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  licenseNumber: string;
  createdAt: string;
}

export interface CreateCustomerRequest {
  firstName: string;
  lastName: string;
  email: string;
  phone: string;
  address: string;
  licenseNumber: string;
}

export interface VehicleType {
  id: number;
  name: string;
  description: string;
  category: string;
}

export interface CreateVehicleTypeRequest {
  name: string;
  description: string;
  category: string;
}

export interface Vehicle {
  id: number;
  registrationNumber: string;
  make: string;
  model: string;
  year: number;
  color: string;
  fuelType: string;
  status: string;
  vehicleTypeId: number;
  vehicleTypeName: string;
  customerId: number;
  customerName: string;
  createdAt: string;
}

export interface CreateVehicleRequest {
  registrationNumber: string;
  make: string;
  model: string;
  year: number;
  color: string;
  fuelType: string;
  status: string;
  vehicleTypeId: number;
  customerId: number;
}