# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vehicle Management System API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: customers
    description: Customer management
  - name: vehicleTypes
    description: Vehicle type management
  - name: vehicles
    description: Vehicle management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Customer deleted

  /vehicle-types:
    get:
      operationId: listVehicleTypes
      tags: [vehicleTypes]
      summary: List all vehicle types
      responses:
        "200":
          description: List of vehicle types
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VehicleType"
    post:
      operationId: createVehicleType
      tags: [vehicleTypes]
      summary: Create a vehicle type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleTypeRequest"
      responses:
        "201":
          description: Vehicle type created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleType"

  /vehicle-types/{id}:
    put:
      operationId: updateVehicleType
      tags: [vehicleTypes]
      summary: Update vehicle type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleTypeRequest"
      responses:
        "200":
          description: Vehicle type updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VehicleType"
    delete:
      operationId: deleteVehicleType
      tags: [vehicleTypes]
      summary: Delete vehicle type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Vehicle type deleted

  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List all vehicles
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Register a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleRequest"
      responses:
        "201":
          description: Vehicle registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"

  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleRequest"
      responses:
        "200":
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Vehicle deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Customer:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        licenseNumber:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - phone
        - address
        - licenseNumber
        - createdAt

    CreateCustomerRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        licenseNumber:
          type: string
      required:
        - firstName
        - lastName
        - email
        - phone
        - address
        - licenseNumber

    VehicleType:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
      required:
        - id
        - name
        - description
        - category

    CreateVehicleTypeRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
      required:
        - name
        - description
        - category

    Vehicle:
      type: object
      properties:
        id:
          type: integer
        registrationNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        fuelType:
          type: string
        status:
          type: string
        vehicleTypeId:
          type: integer
        vehicleTypeName:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - registrationNumber
        - make
        - model
        - year
        - color
        - fuelType
        - status
        - vehicleTypeId
        - vehicleTypeName
        - customerId
        - customerName
        - createdAt

    CreateVehicleRequest:
      type: object
      properties:
        registrationNumber:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        color:
          type: string
        fuelType:
          type: string
        status:
          type: string
        vehicleTypeId:
          type: integer
        customerId:
          type: integer
      required:
        - registrationNumber
        - make
        - model
        - year
        - color
        - fuelType
        - status
        - vehicleTypeId
        - customerId