# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Vehicle {
  id: number;
  registrationNumber: string;
  make: string;
  model: string;
  year: number;
  color: string;
  fuelType: string;
  status: string;
  vehicleTypeId: number;
  vehicleTypeName: string;
  customerId: number;
  customerName: string;
  createdAt: Date;
}