# lib/api-zod/src/generated/types/vehicleType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vehicle Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleType {
  id: number;
  name: string;
  description: string;
  category: string;
}